+++ tasks/seq/karaseva_e_reduce/func_tests/main.cpp
@@ -32,7 +32,7 @@ TEST(karaseva_e_reduce_seq, test_reduce_50) {
   test_task_sequential.Run();
   test_task_sequential.PostProcessing();
 
-  EXPECT_EQ(out[0], static_cast<int>(kCount));
+  EXPECT_EQ(static_cast<size_t>(out[0]), kCount);
 }
 
 TEST(karaseva_e_reduce_seq, test_reduce_100_from_file) {
@@ -66,5 +66,5 @@ TEST(karaseva_e_reduce_seq, test_reduce_100_from_file) {
   test_task_sequential.Run();
   test_task_sequential.PostProcessing();
 
-  EXPECT_EQ(out[0], static_cast<int>(count));
+  EXPECT_EQ(static_cast<size_t>(out[0]), count);
 }
